% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtPQN-package.R
\docType{package}
\name{ProtPQN-package}
\alias{ProtPQN}
\alias{ProtPQN-package}
\title{ProtPQN: Protein-Specific Probabilistic Quotient Normalization for Affinity Proteomics Data From Liquid Biopsies}
\description{
Apply Protein-specific Probabilistic Quotient Normalization (ProtPQN) to affinity proteomics data from liquid biopsies. Use on any wide-format matrix or data frame containing protein measurements, or use a wrapper function to apply on wide or long data with options for transforming back from log-scale data before normalization and normalization in a panel/kit-wise manner. The normalization method is based on Probabilistic Quotient Normalization (Dieterle et al., 2006).
}
\author{
\strong{Maintainer}: Leo Dahl \email{leo.dahl@scilifelab.se}

Authors:
\itemize{
  \item Hampus Hagelin \email{hampus.hagelin@scilifelab.se}
  \item Mun-Gwan Hong
}

Other contributors:
\itemize{
  \item Annika Bendes [contributor]
  \item Jochen Schwenk [contributor]
}

}
\keyword{internal}
