% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protpqn.R
\name{protpqn}
\alias{protpqn}
\title{ProtPQN}
\usage{
protpqn(X)
}
\arguments{
\item{X}{Data frame in wide format containing only protein measurements to be normalized. SampleIDs could e.g. be stored as rownames.}
}
\value{
Output is a data frame or matrix in the same format as input with normalized protein levels.
}
\description{
Perform Protein-specific Probabilistic Quotient Normalization on a matrix.
}
\examples{
# Simple example
protein_data <- replicate(50, rnorm(100, 10, 1))
normalized_data <- protpqn(protein_data)
}
